requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python_dotenv==1.0.0
openai==0.27.6
langchain==0.0.172
google_api_python_client==2.88.0